resources-server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
Flask==2.3.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
PyJWT==2.8.0
MarkupSafe==2.1.3
pycparser==2.21
typing_extensions==4.7.1
uWSGI==2.0.22
Werkzeug==2.3.7
pytest==7.4.1
